Modified do_kafka_test.sh script.

+++ do_kafka_test.sh
@@ -2,12 +2,40 @@
 
 # get name of project folder
 projectFolder=$(pwd | awk -F/ '{print $NF}')
-KAFKA_CONTAINER_NAME="$projectFolder-kafka-1"
-echo "KAFKA_CONTAINER_NAME: $KAFKA_CONTAINER_NAME"
+EXPECTED_KAFKA_CONTAINER_NAME="$projectFolder-kafka-1"
+echo "EXPECTED_KAFKA_CONTAINER_NAME: $EXPECTED_KAFKA_CONTAINER_NAME"
 
 setup() {
-    if [ -z $DOCKER_HOST_IP ]; then
+    # Make sure to set the following environment variables:
+    # - DOCKER_HOST_IP
+    # - DEPOSIT_GROUP
+    # - DEPOSIT_KEY_NAME
+    # - DEPOSIT_BUCKET_NAME
+    # - AWS_REGION
+    # - DEPOSIT_TOPIC
+    # - HEADER_X_API_KEY
+
+    # if any of the above are not set, exit
+    if [ -z "$DOCKER_HOST_IP" ]; then
         echo "DOCKER_HOST_IP is not set. Exiting."
+        exit 1
+    elif [ -z "$DEPOSIT_GROUP" ]; then
+        echo "DEPOSIT_GROUP is not set. Exiting."
+        exit 1
+    elif [ -z "$DEPOSIT_KEY_NAME" ]; then
+        echo "DEPOSIT_KEY_NAME is not set. Exiting."
+        exit 1
+    elif [ -z "$DEPOSIT_BUCKET_NAME" ]; then
+        echo "DEPOSIT_BUCKET_NAME is not set. Exiting."
+        exit 1
+    elif [ -z "$AWS_REGION" ]; then
+        echo "AWS_REGION is not set. Exiting."
+        exit 1
+    elif [ -z "$DEPOSIT_TOPIC" ]; then
+        echo "DEPOSIT_TOPIC is not set. Exiting."
+        exit 1
+    elif [ -z "$HEADER_X_API_KEY" ]; then
+        echo "HEADER_X_API_KEY is not set. Exiting."
         exit 1
     fi
 
@@ -16,13 +44,13 @@ setup() {
 
 waitForKafkaToCreateTopics() {
     while true; do
-        if [ $(docker ps | grep $KAFKA_CONTAINER_NAME | wc -l) == "0" ]; then
-            echo "Kafka container '$KAFKA_CONTAINER_NAME' is not running. Exiting."
+        if [ $(docker ps | grep $EXPECTED_KAFKA_CONTAINER_NAME | wc -l) == "0" ]; then
+            echo "Kafka container '$EXPECTED_KAFKA_CONTAINER_NAME' is not running. Exiting."
             ./stop_kafka.sh
             exit 1
         fi
 
-        ltopics=$(docker exec -it $KAFKA_CONTAINER_NAME /opt/kafka/bin/kafka-topics.sh --list --zookeeper 172.17.0.1)
+        ltopics=$(docker exec -it $EXPECTED_KAFKA_CONTAINER_NAME /opt/kafka/bin/kafka-topics.sh --list --zookeeper 172.17.0.1)
         allTopicsCreated=true
         if [ $(echo $ltopics | grep "topic.OdeBsmJson" | wc -l) == "0" ]; then
             allTopicsCreated=false
@@ -43,6 +71,7 @@ runS3D() {
     ./run.sh &
 
     # wait for S3D to start
+    echo "Sleeping for 10 seconds."
     sleep 10
 
     # use pidof to get pid of S3D
@@ -57,17 +86,19 @@ runS3D() {
 }
 
 pushTestDataToKafka() {
-    echo ""
-    echo "Step 7/$numSteps - Pushing test data to Kafka."
     ./send_data_to_topic.sh
+    echo "Sleeping for 5 seconds."
+    sleep 5
 }
 
 cleanup() {
+    # kill S3D if pid is not empty
+    if [ ! -z "$pid" ]; then
+        echo "Killing S3D with pid $pid."
+        kill $pid
+    fi
+
     ./stop_kafka.sh
-    
-    # kill S3D
-    echo "Killing S3D with pid $pid."
-    kill $pid
 }
 
 run() {
